Add files via upload
I am hesitant to submit an unfinished project, but unfortunately this is all I could produce in the four hours, due to my tiny old computer being unable to read the data files. For example, it took about an hour and a half just to download the price data. I do plan on getting a new computer soon (and my work computer I would've normally used is currently in the office). I would like to add that this is nowhere near the caliber I normally produce.